プレイヤーの弾の動きを定義。一直線のみ

+++ src/Bullet.java
@@ -1,7 +1,8 @@
 public class Bullet extends Entity{
+    //ラジアン
     double angle;
 
-    Bullet(int x,int y,int width,int height,int angle){
+    Bullet(int x,int y,int width,int height,double angle){
         super(x,y,width,height);
         this.angle=angle;
     }
